requirements.txt
Django==1.7b4
Pillow==2.5.1
django-suit==0.2.9
wsgiref==0.1.2
